Pipeline: reorganize code
Put every operation into dedicated source file.
Use common naming convention of files.
