+++ pipeline/foreach_op_test.go
