+++ pipeline/op_spec_test.go
